pyproject.toml
[project]
name = "langchain-chain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-anthropic>=1.3.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "openai>=2.15.0",
    "openi>=3.0.1",
    "python-dotenv>=1.2.1",
    "uvicorn[stadard]>=0.40.0",
]
